Move rollup and storybook config to root package
