+++ package.json
@@ -13,17 +13,42 @@
   },
   "private": true,
   "devDependencies": {
-    "lerna": "^3.20.2",
     "@babel/core": "^7.10.2",
-    "@babel/preset-env": "^7.10.2",
-    "@babel/preset-react": "^7.10.1",
+    "@babel/plugin-proposal-json-strings": "^7.10.1",
     "@babel/plugin-transform-runtime": "^7.10.1",
     "@babel/plugin-transform-strict-mode": "^7.10.1",
-    "@babel/plugin-proposal-json-strings": "^7.10.1",
+    "@babel/preset-env": "^7.10.2",
+    "@babel/preset-react": "^7.10.1",
+    "@rollup/plugin-babel": "^5.0.3",
+    "@rollup/plugin-json": "^4.1.0",
+    "@rollup/plugin-node-resolve": "^8.0.1",
+    "@storybook/addon-docs": "^5.3.19",
+    "@storybook/addon-storysource": "^5.3.19",
+    "@storybook/addon-viewport": "^5.3.19",
+    "@storybook/addons": "^5.3.19",
+    "@storybook/react": "^5.3.19",
+    "@storybook/theming": "^5.3.19",
+    "acorn-jsx": "^5.2.0",
+    "babel-loader": "^8.1.0",
     "babel-plugin-transform-node-env-inline": "^0.4.3",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
+    "css-loader": "^3.5.3",
     "eslint": "^6.6.0",
-    "eslint-config-nebenan": "^1.3.2"
+    "eslint-config-nebenan": "^1.3.2",
+    "lerna": "^3.20.2",
+    "lodash": "^4.17.15",
+    "nebenan-ui-kit": "^4.12.0",
+    "postcss-preset-env": "^6.7.0",
+    "rollup": "^2.15.0",
+    "rollup-plugin-copy": "^3.3.0",
+    "rollup-plugin-delete": "^2.0.0",
+    "rollup-plugin-glob-import": "^0.4.5",
+    "rollup-plugin-peer-deps-external": "^2.2.2",
+    "rollup-plugin-postcss": "^3.1.2",
+    "sass": "^1.26.8",
+    "sass-functions": "^1.4.0",
+    "sass-loader": "^8.0.2",
+    "style-loader": "^1.2.1"
   },
   "scripts": {
     "lint": "lerna run lint",
