--- packages/components/.storybook/theme.js
@@ -1,7 +1,0 @@
-import { create } from '@storybook/theming/create';
-import pkg from "../package.json";
-
-export default create({
-  base: 'light',
-  brandTitle: pkg.name,
-});
