+++ packages/components/package.json
@@ -19,46 +19,23 @@
     "reinstall": "rm -rf node_modules package-lock.json && npm install",
     "prepublishOnly": "npm run check && npm run build",
     "check": "npm run lint && npm run test",
-    "lint": "eslint -f codeframe \"./**/*.{js,jsx}\"",
-    "build": "rollup --config rollup.config.js",
-    "start": "start-storybook -p 3000 --ci",
+    "lint": "npx eslint -f codeframe \"./**/*.{js,jsx}\"",
+    "build": "npx rollup --config rollup.config.js",
+    "start": "npx start-storybook -p 3000 --ci",
     "test": "NODE_ENV=test TZ='GMT' mocha \"./{,!(node_modules)/**}/*.test.{js,jsx}\"",
     "test:tofile": "NODE_ENV=test TZ='GMT' mocha --no-colors -R spec \"./{,!(node_modules)/**}/*.test.{js,jsx}\" 2>&1 | tee test-report.log"
   },
   "devDependencies": {
     "@babel/register": "^7.10.1",
     "@babel/runtime": "^7.10.2",
-    "@rollup/plugin-babel": "^5.0.3",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^8.0.1",
-    "@storybook/addon-docs": "^5.3.19",
-    "@storybook/addon-storysource": "^5.3.19",
-    "@storybook/addon-viewport": "^5.3.19",
-    "@storybook/react": "^5.3.19",
-    "acorn-jsx": "^5.2.0",
-    "babel-loader": "^8.1.0",
     "chai": "^4.2.0",
     "clsx": "^1.1.1",
-    "css-loader": "^3.5.3",
     "date-fns": "^2.14.0",
-    "lodash": "^4.17.15",
     "mocha": "^8.0.1",
     "nebenan-helpers": "^4.2.0",
-    "nebenan-ui-kit": "^4.12.0",
-    "postcss-preset-env": "^6.7.0",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "rollup": "^2.15.0",
-    "rollup-plugin-copy": "^3.3.0",
-    "rollup-plugin-delete": "^2.0.0",
-    "rollup-plugin-glob-import": "^0.4.5",
-    "rollup-plugin-peer-deps-external": "^2.2.2",
-    "rollup-plugin-postcss": "^3.1.2",
-    "sass": "^1.26.8",
-    "sass-functions": "^1.4.0",
-    "sass-loader": "^8.0.2",
-    "style-loader": "^1.2.1"
+    "react-dom": "^16.13.1"
   },
   "dependencies": {},
   "peerDependencies": {
